SGHSS1.1/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
Pillow==11.2.1
python-dotenv==1.1.0
drf-yasg==1.21.10
requests==2.32.3
mysqlclient==2.2.7
setuptools==80.4.0
drf-spectacular[swagger-ui]==0.28.0
